Removed src folder and renamed project.
