+++ .github/workflows/tests.yaml
@@ -38,7 +38,7 @@ jobs:
           uv pip install --system .[dev]
 
       - name: Run tests with pytest
-        run: pytest --cov=chemistry_MEP_TS_optimization --cov-report=xml
+        run: pytest --cov=transbymep --cov-report=xml
 
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v4
