+++ MANIFEST.in
@@ -1,1 +1,1 @@
-graft src/chemistry_MEP_TS_optimization
+graft transbymep
