+++ docs/example_docs/code/source.md
@@ -2,9 +2,9 @@
 
 ## Adding Your Code
 
-All source code (i.e. your various modules, functions, classes, and so on) should be placed in the `/src/<MyPackageName>` directory. A sample file named `examples/sample.py` is included here as a representative example, which you should replace.
+All source code (i.e. your various modules, functions, classes, and so on) should be placed in the `<MyPackageName>` directory. A sample file named `examples/sample.py` is included here as a representative example, which you should replace.
 
-All the code in the `src` directory can be imported now that you have installed your package.
+All the code in the `transbymep` directory can be imported now that you have installed your package.
 
 !!! Tip
 
@@ -19,7 +19,7 @@ All the code in the `src` directory can be imported now that you have installed 
 
 !!! Note
 
-    For any subfolder within `src/<MyPackageName>` containing Python code, you must have an `__init__.py` file, which will tell Python that this is a module you can import.
+    For any subfolder within `<MyPackageName>` containing Python code, you must have an `__init__.py` file, which will tell Python that this is a module you can import.
 
 ## Docstrings
