+++ docs/gen_ref_pages.py
@@ -12,9 +12,9 @@ import mkdocs_gen_files
 
 nav = mkdocs_gen_files.Nav()
 
-for path in sorted(Path("src").rglob("*.py")):
-    module_path = path.relative_to("src").with_suffix("")
-    doc_path = path.relative_to("src").with_suffix(".md")
+for path in sorted(Path("transbymep").rglob("*.py")):
+    module_path = path.relative_to("transbymep").with_suffix("")
+    doc_path = path.relative_to("transbymep").with_suffix(".md")
     full_doc_path = Path("reference", doc_path)
 
     parts = tuple(module_path.parts)
