+++ docs/index.md
@@ -1,6 +1,6 @@
-# chemistry_MEP_TS_Optimization
+# transbymep
 
-Welcome to the documentation for the `chemistry_MEP_TS_optimization` code!
+Welcome to the documentation for the `transbymep` code!
 This python package provides functionality to obtain minimum energy path (MEP)
 connecting two minima (namely, reactant and product for a chemical system)
 which passes through a saddle point (called, the transition state for a chemical system).
