+++ mkdocs.yml
@@ -1,7 +1,7 @@
-site_name: chemistry_MEP_TS_optimization
+site_name: transbymep
 site_author: YourName
 site_description: >-
-  This is the chemistry_MEP_TS_optimization package!
+  This is the transbymep package!
 nav:
   - Home: index.md
   - Code Documentation: reference/ # keep me!
@@ -10,7 +10,7 @@ nav:
       - example_docs/about/conduct.md
       - example_docs/about/license.md
 
-repo_url: https://github.com/khegazy/chemistry_MEP_TS_optimization/
+repo_url: https://github.com/khegazy/transbymep/
 edit_uri: blob/main/docs/
 
 theme:
