+++ transbymep/optimization/configs/gradient_descent_wolfe_schlegel_elastic_band.yaml
