+++ transbymep/optimization/gradient_descent_orig.py
@@ -2,7 +2,7 @@ import jax
 import jax.numpy as jnp
 from functools import partial
 
-from src.tools.logging import logging
+from transbymep.tools.logging import logging
 from . import path_metrics as path_tools
