+++ transbymep/potentials/base_class.py
