+++ transbymep/potentials/configs/wolfe_schlegel.yaml
