+++ transbymep/potentials/wolfe_schlegel.py
